TM-59 Made int values decimal

+++ src/SFA.DAS.EmployerFinance.Web/Orchestrators/TransfersOrchestrator.cs
@@ -59,7 +59,7 @@ namespace SFA.DAS.EmployerFinance.Web.Orchestrators
                     StartingTransferAllowance = accountDetail.Result.StartingTransferAllowance,
                     FinancialYearString = DateTime.UtcNow.ToFinancialYearString(),
                     ActivePledgesTotalAmount = indexTask.Result.ActivePledgesTotalAmount,
-                    PledgesAndTransferConnections = ((int)(accountDetail.Result.StartingTransferAllowance - indexTask.Result.ActivePledgesTotalAmount)),
+                    PledgesAndTransferConnections = (accountDetail.Result.StartingTransferAllowance - indexTask.Result.ActivePledgesTotalAmount),
                     RemainingTransferAllowance = accountDetail.Result.RemainingTransferAllowance
                     HashedAccountID = hashedAccountId
                 }
